refactor : 분리되어 있던 리뷰 서비스 레이어 테스트를 하나로 병합
